Fix Performance Issue of Unform in every-impl
The implementation of unform in every-impl uses count to determine the end of the collection of loops over. In case the collection is a list, count is O(n). Because count is called at every loop iteration, unforming becomes O(n^2).

I found this performance issue while profiling my application with a list of size 500.